Split `VmrVersionFileMerger` into Json and Version.Details classes (#5253)
